requirements.txt
python-telegram-bot==21.6
python-dotenv
nest-asyncio
